Referesh package in Directory
If some commits have not yet been merged,
this could cause package conflicts which,
might explain what is happening here.

It seems as subscript changes are not,
added automatically to high-level app,
settings.

(i.e. When a package is merged to master;
without a branch to reference it can,
not be edited.)

Therefore un-merged commits are not included in this solution,
but can become easily confused with already merged commits.

This can become increasingly difficult as the number of projects,
simultaneously developed increases.

**Be careful not to mistake pre-release updates
with verified project code**